# PERF: groupby.apply on a non-unique, unsorted index